refactor(runtime): split vm_register_ops.rs into cohesive modules (§7-8)
`vm_register_ops.rs` was 2040 lines, well over the 500-line module limit. Split
it into six files along its functional seams, each under 500 lines, with no
behavior change (the set of fn signatures is identical before/after, modulo
visibility):

- `vm_register_ops.rs` (462) — closure/anon-sub/gather creation and the
  closure-capture machinery (`compute_owned_captures`, `capture_closure_env`,
  `capture_upvalues`, `box_captured_lexicals`, `resolve_closure_code`).
- `vm_register_sub_ops.rs` (429) — lambda/block-closure creation and sub/proto/
  token registration (`exec_register_sub_op`, `exec_register_proto_*`,
  `exec_register_token_op`, `register_native_call_sub`).
- `vm_module_ops.rs` (173) — module ops: `use`/`import`/`no`/`need`, `use lib`,
  and var-export registration.
- `vm_var_trait_ops.rs` (383) — variable trait application
  (`exec_apply_var_trait_op`).
- `vm_typedecl_ops.rs` (438) — type-declaration registration ops: enum, class,
  augment, role, subset.
- `vm_scope_ops.rs` (184) — `subtest`/`react`/`whenever` scope execution +
  `find_parameterized_container_parent`.

The closure-capture helpers are bumped `fn`→`pub(super)` (called from the
lambda/block-closure ops now in `vm_register_sub_ops`); `find_parameterized_container_parent`
is bumped `pub(super)` (called from `vm_var_trait_ops`). Each file carries only
the imports it uses (`Compiler` stays with the closure file; `Symbol` dropped
from `vm_var_trait_ops`/`vm_scope_ops`). Twelfth slice of the §7-8 sweep.
